% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{dcmp}
\alias{dcmp}
\title{COM-Poisson probability mass function}
\usage{
dcmp(y, mu, nu, log = FALSE)
}
\arguments{
\item{y}{a positive integer value.}

\item{mu}{the (approximate) mean parameter.}

\item{nu}{the dispersion parameter.}

\item{log}{logical. If \code{TRUE} it returns the logarithm of pmf.}
}
\description{
COM-Poisson probability mass function
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
