% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{ddpo}
\alias{ddpo}
\title{Double Poisson probability mass function}
\usage{
ddpo(y, mu, phi, log = FALSE)
}
\arguments{
\item{y}{a positive integer value.}

\item{mu}{the (approximate) mean parameter.}

\item{phi}{the dispersion parameter.}

\item{log}{logical. If \code{TRUE} it returns the logarithm of pmf.}
}
\description{
Double Poisson probability mass function
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
