% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{ddwe}
\alias{ddwe}
\title{Discrete Weibull probability mass function}
\usage{
ddwe(y, q, rho, log = FALSE)
}
\arguments{
\item{y}{a positive integer value.}

\item{q}{the "location" parameter (\code{0<q<1}).}

\item{rho}{the dispersion parameter.}

\item{log}{logical. If \code{TRUE} it returns the logarithm of pmf.}
}
\description{
Discrete Weibull probability mass function
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
