% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{dgct}
\alias{dgct}
\title{Gamma-count probability mass function}
\usage{
dgct(y, kappa, alpha, log = FALSE)
}
\arguments{
\item{y}{a positive integer value.}

\item{kappa}{the (asymptote) mean parameter.}

\item{alpha}{the dispersion parameter.}

\item{log}{logical. If \code{TRUE} it returns the logarithm of pmf.}
}
\description{
Gamma-count probability mass function
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
