% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{dgpo}
\alias{dgpo}
\title{Generalized Poisson probability mass function}
\usage{
dgpo(y, mu, sigma, log = FALSE)
}
\arguments{
\item{y}{a positive integer value.}

\item{mu}{the mean parameter.}

\item{sigma}{the dispersion parameter.}

\item{log}{logical. If \code{TRUE} it returns the logarithm of pmf.}
}
\description{
Generalized Poisson probability mass function
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
