% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{dptw}
\alias{dptw}
\title{Poisson-Tweedie probability mass function}
\usage{
dptw(y, mu, omega, power, npts = 100L, log = FALSE)
}
\arguments{
\item{y}{a positive integer value.}

\item{mu}{the mean parameter.}

\item{omega}{the dispersion parameter.}

\item{power}{the power parameter.}

\item{npts}{number of points in the gaussian quadrature.}

\item{log}{logical. If \code{TRUE} it returns the logarithm of pmf.}
}
\description{
Poisson-Tweedie probability mass function
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
