% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{equitest}
\alias{equitest}
\title{Likelihood ratio test for equidispersion}
\usage{
equitest(object, ..., heading)
}
\arguments{
\item{object}{an object (or a list of objects) of class
\code{"flexcm"} with a model nested to Poisson (i.e. Com-Poisson,
Gamma-count, Generalized-Poisson, Double Poisson and
Poissson-Tweedie).}

\item{...}{possibly, others fitted models (objects of class
\code{"flexcm"}).}

\item{heading}{logical; if \code{TRUE}, the model formulas are
printed as heading.}
}
\description{
Likelihood ratio test for equidispersion
}
\details{
Note that, when more than one model is passed the models
  \strong{are not compared among themselves}. Here, each model is
  compared to Poisson and them organized in a table.
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
