% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R, R/methods.R, R/predict.R
\name{print.summary.flexcm}
\alias{print.summary.flexcm}
\alias{flexcm-methods}
\alias{print.flexcm}
\alias{logLik.flexcm}
\alias{coef.flexcm}
\alias{vcov.flexcm}
\alias{model.matrix.flexcm}
\alias{fitted.flexcm}
\title{Method for \code{'flexcm'} objects}
\usage{
\method{print}{summary.flexcm}(x, digits = max(3L, getOption("digits") -
  3L), ...)

\method{print}{flexcm}(x, digits = max(3L, getOption("digits") - 3L),
  ...)

\method{logLik}{flexcm}(object, ...)

\method{coef}{flexcm}(object, ...)

\method{vcov}{flexcm}(object, ...)

\method{model.matrix}{flexcm}(object, ...)

\method{fitted}{flexcm}(object, ...)
}
\arguments{
\item{x}{an object of class \code{"flexcm"}.}

\item{digits}{minimal number of _significant_ digits, see
\code{\link[base]{print.default}}.}

\item{...}{Currently not used.}

\item{object}{an object of class \code{"flexcm"}.}
}
\value{
.
}
\description{
Method for \code{'flexcm'} objects
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
