% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{flexcm}
\alias{flexcm}
\title{Fitting flexible count model models}
\usage{
flexcm(formula, data, model, power = NULL, ...)
}
\arguments{
\item{formula}{An object of class "\code{\link[stats]{formula}}"
describe the model for mean.}

\item{data}{An optional data frame containing the variables in the
model. If not found in data, the variables are taken from
\code{environment(formula)}.}

\item{model}{An character indicating the model used. Options are
\code{"compoisson"}, \code{"gammacount"}, \code{"discreteweibull"},
\code{"generalizedpoisson"}, \code{"doublepoisson"}, and
\code{"poissontweedie"}.}

\item{power}{only for Poisson-Tweedie model
(\code{model="poissontweedie"}) a value to be used as fixed power
parameter for Poisson-Tweedie family. Is \code{NULL}, the power
parameter will be estimated.}

\item{...}{Arguments to be used by \code{\link{flexcm_fit}}.}
}
\value{
An object of class \code{flexcm}.
}
\description{
Fitting flexible count model models
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
