% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{flexcm_fit}
\alias{flexcm_fit}
\title{Maximize log-likelihood functions}
\usage{
flexcm_fit(llfun, X, y, start = NULL, method = c("BFGS", "Nelder-Mead",
  "CG", "L-BFGS-B", "SANN"), lower = -Inf, upper = Inf,
  hessian = TRUE, control = list())
}
\arguments{
\item{llfun}{Function thar returns the negative of the log-likelihood
given \code{params}, \code{X} and \code{y} arguments for some count
model.}

\item{X}{Design matrix.}

\item{y}{Vector of observed count data.}

\item{start}{Initial parameters}

\item{method}{Argument passed to \code{\link[stats]{optim}(...)}.}

\item{lower}{Argument passed to \code{\link[stats]{optim}(...)}.}

\item{upper}{Argument passed to \code{\link[stats]{optim}(...)}.}

\item{hessian}{Argument passed to \code{\link[stats]{optim}(...)}.}

\item{control}{Argument passed to \code{\link[stats]{optim}(...)}.}
}
\value{
A list with estimated parameters and hessian matrix.
}
\description{
Maximize log-likelihood functions
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
