% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{llcmp}
\alias{llcmp}
\title{Negative log-likelihood function from COM-Poisson model}
\usage{
llcmp(params, X, y)
}
\arguments{
\item{params}{vector of model parameters.}

\item{X}{Design matrix related to the (approximate) mean parameter
\eqn{\mu = \exp(X \beta)}.}

\item{y}{Vector of observed count data.}
}
\value{
The computed negative log-likelihood function.
}
\description{
TODO: write the likelihood function.
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
