% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{lrtests-flexcm}
\alias{lrtests-flexcm}
\alias{anova.flexcm}
\alias{anova.list}
\alias{lrtest}
\title{Likelihood ratio tests for nested flexible count models}
\usage{
\method{anova}{flexcm}(object, ..., heading)

\method{anova}{list}(object, ..., heading)

lrtest(object, ..., heading)
}
\arguments{
\item{object}{an object of class \code{flexcm}, a result of call
\code{\link{flexcm}(...)}.}

\item{...}{Currently not used.}

\item{heading}{logical; if \code{TRUE}, the model formulas are
printed as heading. Default is \code{TRUE}.}
}
\value{
an object of class \code{"anova"}, a table with compenents
  for LR test.
}
\description{
Likelihood ratio tests for nested flexible count models
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
