% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.flexcm}
\alias{predict.flexcm}
\title{Predict method for flexible count models}
\usage{
\method{predict}{flexcm}(object, newdata, newmatrix = NULL,
  type = c("link", "response"), interval = c("none", "confidence"),
  level = 0.95, augment_data = FALSE, ...)
}
\arguments{
\item{object}{a fitted object of class from \code{"flexcm"}.}

\item{newdata}{optionally, a data frame in which to look for
variables with which to predict. If omitted, the fitted linear
predictors are used.}

\item{newmatrix}{optionally, a list with named design matrices
(\code{"X"}) used to predict. If omitted, the fitted linear
predictors are used.}

\item{type}{the type of prediction required. The default
\code{"link"} is on the scale of the linear predictors; the
alternative \code{"response"} is on the scale of the response
variable (for the models that the location parameter not is the
mean, it will be computed the mean by \eqn{\sum x p(x; \theta)}).}

\item{interval}{character indicating the type of interval
calculation. Currently the intervals can be \code{"none"} or
\code{"confidence"}.}

\item{level}{level of confidence intervals.}

\item{augment_data}{logical indicating if \code{newdata} should be
augmented with the predict values and, possibly, confidence
intervals. Default is \code{FALSE}.}

\item{...}{currently not used.}
}
\value{
a tibble with fitted values (column code{fit}) and, possibly,
  confidence intervals (columns \code{lwr} and \code{upr}). If
  \code{augment_data}, the result will be the \code{newdata} with new
  columns \code{fit}, fitted values, and code{ste}, standard errors.
}
\description{
Predict method for flexible count models
}
\details{
TODO
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
