% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{summary.flexcm}
\alias{summary.flexcm}
\title{Summary of the flexible count models (individual t-tests)}
\usage{
\method{summary}{flexcm}(object, correlation = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{flexcm}, a result of call
\code{\link{flexcm}(...)}.}

\item{correlation}{logical; if \code{TRUE}, the correlation matrix of
the estimated parameters is returned and printed. Default is
\code{FALSE}.}

\item{...}{Currently not used.}
}
\value{
an object of class \code{"summary.flexcm"}, a list with
  components.
}
\description{
Summary of the flexible count models (individual t-tests)
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
